API Reference added to the TOC
Also updated Makefile to remove modules.rst after running sphinx-apidoc.
The modules.rst file is usually used in projects with a complex module
hierarchy, and is generated automatically with sphinx-apidoc.
In this project it is not used, which causes a sphinx warning during the
build process. (The warning is harmless, but still.)